taskmaster-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.0.3
httpx==0.25.0
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
twilio==8.10.0
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
alembic==1.13.0
email-validator==2.1.0
